feat(auth): create and implement userAuthStore
When the user login with Google, their username and profile photo need to appear on the account button in the header/sidebar.

To do this, created useUserAuthStore with Pinia using TypeScript.
This was implemented in the login page script to capture the information (displayName and photoURL) when authenticating with Vuefire and also in the AccountButton component, which causes the state to update if the user is logged in and shows the information in question.